requirements.txt
open cv
numpy
face_recognition
imutils
mysql.connector
jason
dlib
requests
os
tkinter
